tabular-template/requirements.txt
alembic==1.4.2
appnope==0.1.0
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
bleach==3.1.3
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
configparser==5.0.0
cycler==0.10.0
databricks-cli==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docker==4.2.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
gitdb==4.0.5
GitPython==3.1.2
gorilla==0.3.0
gunicorn==20.0.4
hydra-core==0.11.3
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
lightgbm==2.3.1
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mlflow==1.8.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.2
omegaconf==1.4.1
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.5.4
prometheus-client==0.7.1
prometheus-flask-exporter==0.13.0
prompt-toolkit==3.0.4
protobuf==3.11.3
ptyprocess==0.6.0
pyarrow==0.16.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
querystring-parser==1.2.4
requests==2.23.0
retrying==1.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.14.0
sklearn==0.0
smmap==3.0.4
SQLAlchemy==1.3.13
sqlparse==0.3.1
tabulate==0.8.7
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.1.0
